pyproject.toml
[project]
name = "romanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "django>=6.0",
    "django-allauth>=65.14.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "paramiko>=4.0.0",
    "pillow>=12.0.0",
    "procrastinate>=3.6.0",
    "psycopg[binary]>=3.3.2",
    "py7zr>=1.0.0",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.8",
    "pre-commit>=4.0.0",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "romhoard.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
